% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CER.R
\name{CER}
\alias{CER}
\title{Classification Error Rate (CER)}
\usage{
CER(reference, estimate)
}
\arguments{
\item{reference}{the vector containing the reference classification.}

\item{estimate}{the vector containing the estimated classification.}
}
\value{
The function returns the CER. The closer is the value to 0,
the better is the estimated classification with respect to the reference one.
}
\description{
This function returns the classification error rate (CER, *Witten and Tibshirani, 2010*) as a measure of the discrepancy between a reference classification and an estimated classification.
}
\examples{
a <- c(1, 0, 0, 1, 1)
b <- c(1, 1, 0, 1, 1)
CER(a, b)
}
\references{
Witten, D. M. and Tibshirani, R. (2010). A Framework for Feature Selection in Clustering. *Journal of the American Statistical Association* **105** 713-726.
}
