% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSpartacus.R
\name{combineSpartacus}
\alias{combineSpartacus}
\title{Manually combine multiple runs of SpaRTaCUS}
\usage{
combineSpartacus(x = NULL, search.dir = NULL, compute.uncertainty = T)
}
\arguments{
\item{x}{either a list of `spartacus` objects or a vector containing the names of the files to be combined;}

\item{search.dir}{the directory path where the file names given by `x` are searched. If `NULL`, the file paths must be specified as part of the file names is `x`;}

\item{compute.uncertainty}{if `TRUE` (default), it computes the clustering uncertainty of the rows.}
}
\value{
an object of class `spartacus` with the parameter estimates and the clustering labels given by the best fit across the ones in `x`. If `compute.uncertainty == T`, the clustering uncertainty measures are returned.
}
\description{
This function combines multiple runs of SpaRTaCUS with the same values of K and R.
}
\details{
If each element of `x` is an object of class `spartacus`, then the function combines them into a unique spartaco object.
If the `spartaco` objects are stored into different files, the vector `x` receives the names of the files,
and `search.dir` is the path of the directory where the files are stored.
}
