% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneVariances.approx.R
\name{geneVariances.approx}
\alias{geneVariances.approx}
\title{Compute the gene-specific variances}
\usage{
geneVariances.approx(
  x,
  gene.names = NULL,
  MCMC.interval = list(level = 0.95, R = 10^4)
)
}
\arguments{
\item{x}{an object of class `spartaco`;}

\item{gene.names}{a vector of length `nrow(x$x)`, containing the gene names. If `NULL`, the function takes `gene.names = row.names(x$x)`.}

\item{MCMC.interval}{a list of length 2 containing the options for computing the high posterior density intervals of the gene specific variances. The  level is given by `MCMC.interval$level` (default `0.95`).
Intervals are computed via MCMC; the number of simulated draws must is passed to `MCMC.interval$level` (default is `10^4`).
If `NULL`, the intervals are computed as mean + c(-1.96,1.96)*sd, and the function returns a warning.}
}
\value{
An object of class `spartacus.genes` containing five elements:
- `Expectation`: a data frame of dimension `nrow(x$x)` x `R` containing the expected values of the random effects within the `R` spot clusters;
- `Variance`: a data frame of dimension `nrow(x$x)` x `R` containing the variance of the random effects within the `R` spot clusters;
- `HPD.left`: a data frame of dimension `nrow(x$x)` x `R` containing the left HPD interval of the random effects within the `R` spot clusters;
- `HPD.right` a data frame of dimension `nrow(x$x)` x `R` containing the right HPD interval of the random effects within the `R` spot clusters;
- `Cs`: the row clustering labels (the same contained into the object `x`).
- `Alpha.new`: the updated estiamte of the Alpha parameter of the SpaRTaCUS model.
- `Beta.new`: the updated estiamte of the Beta parameter of the SpaRTaCUS model.
}
\description{
This function returns the summary statistics for the random variances estimated through SpaRTaCUS within the `K*R` blocks.
}
