% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spartacus.R
\name{plot.spartacus}
\alias{plot.spartacus}
\title{Plot SpaRTaCo
This function returns the ggplots of the mean and the spatial signal-to-noise ratios from a SpaRTaCUS model.}
\usage{
\method{plot}{spartacus}(
  x,
  type = 1,
  gene.name = readline("gene name: "),
  k = NULL,
  r = 1:ncol(x$mu),
  manual.palette = NULL,
  display.all.spots = T,
  range.mu = NULL,
  range.sRatio = NULL,
  size = NULL,
  ...
)
}
\arguments{
\item{x}{a `spartacus` object;}

\item{type}{the type of plot you want to show.
- `1` displays the co-cluster mean levels;
- `2` displays the co-cluster spatial signal-to-noise ratios;
- `3` displays the map of the spots colored with respect to the estimated column clusters;
- `4` if `!is.null(k)`, it displays the average gene expression in the clusters `k` and `r`, otherwise it displays the expression of the gene given in `gene.name`.}

\item{gene.name}{(used only when `type == 4`); it receives the name of the gene to display. If `!is.null(k)`, it displays the average expression of the gene clusters given by `k`.}

\item{k}{(used only when `type == 4`) the gene clusters to plot.}

\item{r}{(used only when `type == 4`) the spot clusters to plot.}

\item{manual.palette}{a vector of colors used when `type == 3`.}

\item{display.all.spots}{if `TRUE` (default) and `type == 4`, it displays the entire grid of spots.}

\item{range.mu}{set min and max values of mean's plot scale.}

\item{range.sRatio}{set min and max values of spatian signal to noise Ratio's plot scale.}

\item{...}{}
}
\value{
The requested plot is displayed. In addition, if assigned to an object, it will return the `ggplot` object.
}
\description{
Plot SpaRTaCo
This function returns the ggplots of the mean and the spatial signal-to-noise ratios from a SpaRTaCUS model.
}
