% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartacus.internal.R
\name{spartacus.internal}
\alias{spartacus.internal}
\title{SpaRTaCUS with Column Cluster assigned}
\usage{
spartacus.internal(
  data,
  assay = NULL,
  coordinates = NULL,
  K = K,
  R = R,
  unsupervised,
  spot.labels,
  n.neighbors = 20,
  Alpha = 10,
  Tau = 1,
  input.values = NULL,
  max.iter = 1000,
  conv.criterion = list(iterations = 10, epsilon = 1e-04),
  verbose = F,
  save.options = NULL,
  seed = NULL
)
}
\arguments{
\item{data}{either a `SpatialExperiment` object or a matrix containing the experiment.}

\item{assay}{if `class(data) == "SpatialExperiment"`, it takes either the name or the index of the assay to be used.}

\item{coordinates}{if `is.matrix(data)`, it takes the matrix of spatial coordinates of dimension `ncol(data)` x 2.}

\item{K}{the number of row clusters (only when `input.values == NULL`);}

\item{R}{numeber of spot cluster}

\item{unsupervised}{if true unspervised clustering is performed}

\item{spot.labels}{a vector containing the labels of column cluster.}

\item{n.neighbors}{the number of nearest neighbors for fitting the nearest-neighbor Gaussian process (NNGP) model. The default value is 20. ??Da aggiungere commento sulla scelta di m??}

\item{Alpha}{the constraint for the Alpha parameters. ??Da aggiungere nei dettagli le caratteristiche del vincolo, se lasciamo la libertà di farlo??}

\item{Tau}{the constraint for the Tau parameters. ??Da aggiungere nei dettagli le caratteristiche del vincolo, se lasciamo la libertà di farlo??}

\item{input.values}{starting value for parameter estimation}

\item{max.iter}{the maximum number of iterations the estimation algorithm is run.}

\item{conv.criterion}{a list containing the parameters that define a converge criterion (see **Details**).}

\item{verbose}{progress of estimation}

\item{save.options}{a list for specifying the saving parameters (see **Details**).}

\item{seed}{set the internal seed of the function.}
}
\value{
An object of class `spartacus` with the parameter estimates, the clustering labels, the log-likelihood value at each iteration and the data, the ICL, the data matrix and the coordinates matrix.
}
\description{
This function returns the estimated model parameters and the row-clustering labels.
}
\details{
The algorithm can be initiated from a given set of starting values. To do so, `input.values` receives a list of the form
`list(mu, tau, xi, alpha, beta, phi, Cs)`, where:
- `mu`, `delta` and `beta` are `K` x `R` matrices;
- `phi` is a vector of length `R`;
- `Cs` is a vector of length `nrow(data)` containing the row clustering labels;

If the algorithm is initiated from some starting values,  `K` is set automatically according to the input values.
If an object of class `spartacus` is passed to `input.values`, the estimation starts from the final estimate of the previous run (see **Examples**).

If `conv.criterion == NULL`, the algorithm is stopped after `max.iter` itereations, otherwise it is stopped when the increment of the log-likelihood is smaller than a certain threshold `conv.criterion$epsilon` for `conv.criterion$iterations` times in a row.

The function allows also to save the results even if the estimation is not completed. To do so, `save.options` receives a list of two parameters:
`after` gives the number of iterations after which the results are saved, `file.name` contains the path where the results are saved.

If `verbose == "full"`, the on-going estimation procedure is displayed. If `verbose == "progress"`, a dynamic progress bar will display the percentage of iterations completed.
If `verbose == F`, then nothing is displayed in console.
}
