% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAlphaBeta.R
\name{updateAlphaBeta}
\alias{updateAlphaBeta}
\title{Update Alpha and Beta estimation}
\usage{
updateAlphaBeta(x, n.neighbors = 20, gene.names = NULL)
}
\arguments{
\item{x}{an object of class `spartacus`.}

\item{n.neighbors}{the number of nearest neighbors for fitting the nearest-neighbor Gaussian process (NNGP) model. The default value is 20. ??Ha senso lasciare la libertà di impostarlo, forse no??}

\item{gene.names}{a vector of length `nrow(x$x)`, containing the gene names. If `NULL`, the function takes `gene.names = row.names(x$x)`.}
}
\value{
a `spartacus` object with the updated estimate of Alpha and Beta
}
\description{
Update Alpha and Beta estimation
}
